Реализована система проверки прав пользователя

+++ app/repo/Delete.php
@@ -1,16 +1,21 @@
 <?php 
+if ($row['perm'] == 'admin')
+{
+    $id = $_GET["id"] ?? "";
+    $type = $_GET["type"] ?? "";
 
-$id = $_GET["id"] ?? "";
-$type = $_GET["type"] ?? "";
+    if (!isset($tables[$type])) {
+        die("Неверный тип!");
+    }
 
-if (!isset($tables[$type])) {
-    die("Неверный тип!");
+    $database = $tables[$type]; 
+
+    $sql = "DELETE FROM $database WHERE id = '$id'";
+    $row = $pdo->query($sql);
+    header("Location: ". $_SERVER['HTTP_REFERER']);
 }
-
-$database = $tables[$type]; 
-
-$sql = "DELETE FROM $database WHERE id = '$id'";
-$row = $pdo->query($sql);
-header("Location: ".$_SERVER['HTTP_REFERER']);
-
+else
+{
+    echo '<h1> Недостаточно прав! </h1>';     
+}
